requirements.txt
pandas>=2.2.0
SQLAlchemy>=2.0.0
requests>=2.32.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
APScheduler>=3.10.0
streamlit>=1.35.0
python-dotenv>=1.0.0
